# .well-known/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "GPT Plugin Diagrams",
        "version": "1.1.1"
    },
    "servers": [
        {
            "url": "http://localhost:5003"
        }
    ],
    "paths": {
...
        "/generate_diagram": {
            "post": {
                "operationId": "generate_diagram_endpoint",
                "parameters": [
                    {
                        "in": "query",
                        "name": "lang",
                        "required": true,
                        "schema": {
                            "title": "Language",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "diagram_type",
                        "required": true,
                        "schema": {
                            "title": "Diagram Type",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "diagram_text",
                        "required": true,
                        "schema": {
                            "title": "Text",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Generate Diagram Endpoint",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Generate Diagram Endpoint"
            }
        },
...
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "title": "Detail",
                        "type": "array"
                    }
                },
                "title": "HTTPValidationError",
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "title": "Location",
                        "type": "array"
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                },
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError",
                "type": "object"
            }
        }
    }
}